# devops-final-be/docs/swagger.json
{
    "schemes": [
        "https",
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Backend API for the DevOps demo project.",
        "title": "DevOps Demo API",
        "contact": {}
    },
    "basePath": "/",
    "paths": {
...
        "/api/warehouse/dashboard/summary": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get warehouse dashboard summary including low stock items and pending approval requests",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get Warehouse Dashboard Summary",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string"
                                },
                                "result": {
                                    "$ref": "#/definitions/models.WarehouseDashboardSummaryResponse"
                                },
                                "status": {
                                    "type": "string"
                                },
                                "status_code": {
                                    "type": "integer"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string"
                                },
                                "result": {},
                                "status": {
                                    "type": "string"
                                },
                                "status_code": {
                                    "type": "integer"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string"
                                },
                                "result": {},
                                "status": {
                                    "type": "string"
                                },
                                "status_code": {
                                    "type": "integer"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/warehouse/items": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get warehouse items with optional search and category filters",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Warehouse"
                ],
                "summary": "Get Warehouse Items",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search by item name or code",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Category (MED, EQU, CON)",
                        "name": "category",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string"
                                },
                                "result": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                },
                                "status": {
                                    "type": "string"
                                },
                                "status_code": {
                                    "type": "integer"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string"
                                },
                                "result": {},
                                "status": {
                                    "type": "string"
                                },
                                "status_code": {
                                    "type": "integer"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string"
                                },
                                "result": {},
                                "status": {
                                    "type": "string"
                                },
                                "status_code": {
                                    "type": "integer"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string"
                                },
                                "result": {},
                                "status": {
                                    "type": "string"
                                },
                                "status_code": {
                                    "type": "integer"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new warehouse item",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Warehouse"
                ],
                "summary": "Create Warehouse Item",
                "parameters": [
                    {
                        "description": "Warehouse item payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateWarehouseItemRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string"
                                },
                                "result": {
                                    "type": "object"
                                },
                                "status": {
                                    "type": "string"
                                },
                                "status_code": {
                                    "type": "integer"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string"
                                },
                                "result": {},
                                "status": {
                                    "type": "string"
                                },
                                "status_code": {
                                    "type": "integer"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string"
                                },
                                "result": {},
                                "status": {
                                    "type": "string"
                                },
                                "status_code": {
                                    "type": "integer"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string"
                                },
                                "result": {},
                                "status": {
                                    "type": "string"
                                },
                                "status_code": {
                                    "type": "integer"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.CreateWarehouseItemRequest": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "code": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "minimumQuantity": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                },
                "unit": {
                    "type": "string"
                }
            }
        },
...
        "models.WarehouseDashboardSummaryResponse": {
            "type": "object",
            "properties": {
                "low_stock_items_count": {
                    "type": "integer"
                },
                "low_stock_threshold": {
                    "type": "integer"
                },
                "pending_restock_requests_count": {
                    "type": "integer"
                },
                "pending_withdraw_requests_count": {
                    "type": "integer"
                },
                "total_items_count": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}